pyproject.toml
[tool.poetry]
name = "loan-default-prediction"
version = "0.1.0"
description = "MLOps package for loan default prediction using Transformer models"
authors = ["WB-Jang"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.0"
apache-airflow = "^2.7.0"
apache-airflow-providers-postgres = "^5.7.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
mlflow = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
